requirements.txt
# Core LLM & HTTP
openai>=1.0.0
anthropic>=0.40.0
google-genai>=0.8.0
requests>=2.28.0
httpx>=0.25.0
python-dotenv>=1.0.0
tqdm>=4.65.0
certifi
holidays>=0.89

# Documents & Data
pdfplumber>=0.9.0
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.0.0
beautifulsoup4>=4.11.0
exstruct>=0.3.3
pyyaml>=6.0
numpy

# Network & Browser
pyOpenSSL
playwright

# GUI
pyside6

# Windows / OS Specific
pyreadline3>=3.4.1; platform_system == "Windows"
pyautogui; platform_system == "Windows" or platform_system == "Darwin" or platform_system == "Linux"
pygetwindow; platform_system == "Windows"

# Dev Tools (Optional)
mdformat>=0.7.0
ruff>=0.0.0
black>=24.0.0
mypy>=1.0.0
pylint>=3.0.0
